Remove reciprocal resets until we can fix hooks error
We have a hooks error when hooks are not called in the same order
across renders, which happens when a new (or just the first?) object
is created, immediately after adding a vnum and tabbing out of that
field.

Until we can fix it, just remove this component. It's purely a
convenience.